Feat:Quiz validations
