+++ src/main/kotlin/com/evalify/evalifybackend/quiz/domain/DTO/QuizStudentDTO.kt
@@ -1,0 +1,5 @@
+package com.evalify.evalifybackend.quiz.domain.DTO
+
+class QuizStudentDTO (
+    val ipAddress:String
+    ){}
