+++ src/main/kotlin/com/evalify/evalifybackend/quiz/question/domain/DescriptiveQuestion.kt
@@ -1,20 +1,18 @@
-package com.evalify.evalifybackend.question.domain
+package com.evalify.evalifybackend.quiz.question.domain
 
 import com.evalify.evalifybackend.bank.domain.Bank
 import com.evalify.evalifybackend.questions.domain.BaseQuestion
 import com.evalify.evalifybackend.questions.domain.Difficulty
 import com.evalify.evalifybackend.questions.domain.QuestionTypes
 import com.evalify.evalifybackend.questions.domain.Taxonomy
-import com.vladmihalcea.hibernate.type.json.JsonBinaryType
-import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
-import org.hibernate.annotations.Type
 import java.util.UUID
 
+
 @Entity
-@DiscriminatorValue(value = "CODING")
-class CodingQuestion(
+@DiscriminatorValue(value = "DESCRIPTIVE")
+class DescriptiveQuestion(
     id: UUID?,
     question: String = "",
     bank: Bank?,
@@ -26,19 +24,10 @@ class CodingQuestion(
     co: Int,
     negativeMark: Int? = null,
     difficulty: Difficulty,
-    val driverCode: String?,
-    val boilerCode: String?,
-    val functionName: String?,
-    val returnType: String?,
-    @Type(JsonBinaryType::class)
-    @Column(columnDefinition = "jsonb")
-    val params: List<FunctionParam>?,
-    @Type(JsonBinaryType::class)
-    @Column(columnDefinition = "jsonb")
-    val testcases: List<TestCase>?,
-    val language: List<String>?,
-    val answer: String?
-): BaseQuestion(
+    val expectedAnswer:String?,
+    val strictness:Float?,
+    val guidelines:String?
+) : BaseQuestion(
     id=id,
     question = question,
     bank = bank,
@@ -51,8 +40,8 @@ class CodingQuestion(
     negativeMark = negativeMark,
     difficulty = difficulty
 ){
-    override fun copyQuestion(): CodingQuestion {
-        val copiedQuestion = CodingQuestion(
+    override fun copyQuestion(): DescriptiveQuestion {
+        val copiedQuestion = DescriptiveQuestion(
             id = null,
             question = question,
             bank = bank,
@@ -64,19 +53,11 @@ class CodingQuestion(
             co = co,
             negativeMark = negativeMark,
             difficulty = difficulty,
-            driverCode = driverCode,
-            boilerCode = boilerCode,
-            functionName = functionName,
-            returnType = returnType,
-            params = params,
-            testcases = testcases,
-            language = language,
-            answer = answer
+            expectedAnswer = expectedAnswer,
+            strictness = strictness,
+            guidelines = guidelines,
+
         )
         return copiedQuestion
     }
 }
-
- class FunctionParam(val param: String, val type: String)
-
- class TestCase(val input: List<Any>, val expected: Any)
