+++ src/main/kotlin/com/evalify/evalifybackend/quiz/question/domain/MatchTheFollowing.kt
@@ -1,33 +1,38 @@
-package com.evalify.evalifybackend.question.domain
+package com.evalify.evalifybackend.quiz.question.domain
 
 import com.evalify.evalifybackend.bank.domain.Bank
 import com.evalify.evalifybackend.questions.domain.BaseQuestion
 import com.evalify.evalifybackend.questions.domain.Difficulty
 import com.evalify.evalifybackend.questions.domain.QuestionTypes
 import com.evalify.evalifybackend.questions.domain.Taxonomy
+import com.vladmihalcea.hibernate.type.json.JsonBinaryType
+import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
+import org.hibernate.annotations.Type
 import java.util.UUID
 
+class MatchPair(val id: String, val leftPair: String,val rightPair:String)
 
 @Entity
-@DiscriminatorValue(value = "DESCRIPTIVE")
-class DescriptiveQuestion(
+@DiscriminatorValue(value = "MATCH_THE_FOLLOWING")
+
+class MatchTheFollowing (
     id: UUID?,
     question: String = "",
     bank: Bank?,
 //    topic: MutableList<Topic>,
     explanation: String? = "",
-    hint: String? = "",
-    marks: Int,
+    hint: String? = "", marks: Int,
     bloomsTaxonomy: Taxonomy,
     co: Int,
     negativeMark: Int? = null,
     difficulty: Difficulty,
-    val expectedAnswer:String?,
-    val strictness:Float?,
-    val guidelines:String?
-) : BaseQuestion(
+    @Type(JsonBinaryType::class)
+    @Column(columnDefinition = "jsonb")
+    val keys:MutableList<MatchPair> = mutableListOf<MatchPair>()
+
+): BaseQuestion(
     id=id,
     question = question,
     bank = bank,
@@ -39,9 +44,9 @@ class DescriptiveQuestion(
     co = co,
     negativeMark = negativeMark,
     difficulty = difficulty
-){
-    override fun copyQuestion(): DescriptiveQuestion {
-        val copiedQuestion = DescriptiveQuestion(
+) {
+    override fun copyQuestion(): MatchTheFollowing {
+        val copiedQuestion = MatchTheFollowing(
             id = null,
             question = question,
             bank = bank,
@@ -53,10 +58,7 @@ class DescriptiveQuestion(
             co = co,
             negativeMark = negativeMark,
             difficulty = difficulty,
-            expectedAnswer = expectedAnswer,
-            strictness = strictness,
-            guidelines = guidelines,
-
+            keys = keys
         )
         return copiedQuestion
     }
