+++ src/main/kotlin/com/evalify/evalifybackend/quiz/question/repository/BankQuestionRepository.kt
@@ -1,6 +1,6 @@
-package com.evalify.evalifybackend.question.repository
+package com.evalify.evalifybackend.quiz.question.repository
 
-import com.evalify.evalifybackend.question.domain.bankQuestion.BankQuestion
+import com.evalify.evalifybackend.quiz.question.domain.bankQuestion.BankQuestion
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
