+++ src/main/kotlin/com/evalify/evalifybackend/quiz/question/repository/QuizQuestionRepository.kt
@@ -1,6 +1,6 @@
-package com.evalify.evalifybackend.question.repository
+package com.evalify.evalifybackend.quiz.question.repository
 
-import com.evalify.evalifybackend.question.domain.quizQuestion.QuizQuestion
+import com.evalify.evalifybackend.quiz.question.domain.quizQuestion.QuizQuestion
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.rest.core.annotation.RepositoryRestResource
 import java.util.UUID
