+++ src/main/kotlin/com/evalify/evalifybackend/section/domain/Section.kt
@@ -1,7 +1,7 @@
 package com.evalify.evalifybackend.section.domain
 
-import com.evalify.evalifybackend.question.domain.quizQuestion.QuizQuestion
 import com.evalify.evalifybackend.quiz.domain.Quiz
+import com.evalify.evalifybackend.quiz.question.domain.quizQuestion.QuizQuestion
 import jakarta.persistence.*
 import java.util.UUID
